Solve exercise 2
Set padding: 0px and margin: 0px on ul, and padding: 8px on a,
to eliminate preset list paddings

Set display: flex on .header and ul, to line up all
items horizontally.

Set display: flex on .left-links and .right-links
and set margins appripriately, in order to center
the logo and keep the links on the edge of the
viewport.

set flex: 0 1 230px on .left-links, .right-links,
and ul to give the list elements adequate spacing.

set flex: 0 0 content on li to keep the list
elements from growing to fit the flex container.

set justify-content: space-around on ul,
in order to provide even spacing between flex
items as well as minor spacing on the edges.