Solve exercise 03
Create .container divs to group .logo and .links,
as well as .notifications and .profile-image.

Set desiplay: flex and align-items: center on
.header, .container, and ul, to line everything
up horizontally on the viewport.

Set gap: 16px for .container and ul, to give
adequate spacing between items.

Set padding: 8px and justify-content:
space-between on .header, to send both containers
to the edges of the viewport.

Set margin and padding to 0px on ul, to avoid
preset margins and paddings on lists and list
elements.