Solve exercise 4
Add containers to create info-boxes and an info-row,
in order to have the right structure for flex boxes.

Add text-align: center and margin-bottom: 25px
to .title for spacing.

Line up items in the info-row using flex. align items
to top of container because of differing text sizes
when changing sizes.

Set the info boxes to display: flex in the column
direction, in order to have images centered. Set
flex behavior to allow for shrinking and a basis
of 200px.

Center-align .text items.